requirements.txt
python-telegram-bot[ext]
google-generativeai
requests
gtts
pytz 